objects/earthquake: merge to trx
