+++ src/libtrx/meson.build
@@ -337,6 +337,7 @@ sources = [
   'game/objects/general/ding_dong.c',
   'game/objects/general/door.c',
   'game/objects/general/drawbridge.c',
+  'game/objects/general/earthquake.c',
   'game/objects/general/general.c',
   'game/objects/general/gong.c',
   'game/objects/general/gong_bonger.c',
